refs #31495 muon analysis clean up of plotting
